second draft

+++ decisiontree/tixtactoegame.h
@@ -16,13 +16,12 @@ public:
 	char winner();
 	void print();
 	void valid();
-
 	int validmoves();
 	int validmovex(int i);
 	int validmovey(int i);
 	char getside();
 private:
-	int size;
+	short int size;
 	char board[10][10];
 	char moves[100][3];
 	vector<int> x;
